requirements.txt
agno>=2.4.0
streamlit>=1.54.0
duckdb>=1.4.0
pandas>=2.0.0
openpyxl>=3.1.0
pyarrow>=14.0.0
python-dotenv>=1.0.0
openai<1.0.0,>=0.28.0  # Agno requires OpenAI client < 1.0.0

# Required for API communication
requests>=2.31.0

# Optional dependencies for enhanced functionality
# altair>=5.0.0  # For advanced visualizations
# plotly>=5.0.0  # For interactive charts
# seaborn>=0.12.0  # For statistical visualizations
# scikit-learn>=1.3.0  # For machine learning analysis
# scipy>=1.11.0  # For scientific computing
# matplotlib>=3.7.0  # For custom plotting